Some rearranging to get legacy to build
