+++ libcommon/include/vcc/util/fileutil.h
@@ -25,6 +25,7 @@
 #include <filesystem>
 #include <algorithm>
 #include <Windows.h>
+//#include <vcc/util/textutil.h>
 
 //=========================================================================
 // Host file utilities.  Most of these are general purpose
@@ -163,98 +164,6 @@ namespace VCC::Util {
 		std::filesystem::path p(input);
 		std::string out = p.parent_path().string();
 		FixDirSlashes(out);
-		return out;
-	}
-
-	//------------------------------------------------------------------------
-	// TODO: Functions that should go in a more general purpose compile unit
-	//------------------------------------------------------------------------
-
-	// Return string with case conversion
-	inline std::string to_lower(std::string s) {
-		std::transform(s.begin(), s.end(), s.begin(),
-			[](unsigned char c) {
-				return static_cast<char>(std::tolower(c));
-			});
-		return s;
-	}	
-
-	inline std::string to_upper(std::string s) {
-		std::transform(s.begin(), s.end(), s.begin(),
-			[](unsigned char c) {
-				return static_cast<char>(std::toupper(c));
-			});
-		return s;
-	}
-
-	inline void make_lower(std::string& s) {
-		std::transform(s.begin(), s.end(), s.begin(),
-			[](unsigned char c) {
-				return static_cast<char>(std::tolower(c));
-			});
-	}
-
-	inline void make_lower(char* s) {
-		if (!s) return;
-		for (char* p = s; *p; ++p)
-			*p = static_cast<char>(std::tolower(static_cast<unsigned char>(*p)));
-	}
-
-	inline void make_upper(std::string& s) {
-		std::transform(s.begin(), s.end(), s.begin(),
-			[](unsigned char c) {
-				return static_cast<char>(std::toupper(c));
-			});
-	}
-
-	inline void make_upper(char* s) {
-		if (!s) return;
-		for (char* p = s; *p; ++p)
-			*p = static_cast<char>(std::toupper(static_cast<unsigned char>(*p)));
-	}
-
-	inline bool is_null_or_empty(const char* s) {
-		return s == nullptr || *s == '\0';
-	}
-
-	inline bool is_null_or_empty(const std::string& s) {
-		return s.empty();
-	}
-
-	inline void copy_to_char(const std::string& src, char* dst, size_t dst_size)
-	{
-		if (dst_size == 0) return;
-		const size_t n = std::min(src.size(), dst_size - 1);
-		memcpy(dst, src.data(), n);
-		dst[n] = '\0';
-	}
-
-	inline std::string WideToUTF8(const wchar_t* w)
-	{
-		if (!w) return {};
-
-		int len = WideCharToMultiByte(CP_UTF8, 0, w, -1, nullptr, 0, nullptr, nullptr);
-		if (len <= 1) return {};
-
-		std::string out(len - 1, 0);
-		WideCharToMultiByte(CP_UTF8, 0, w, -1, out.data(), len, nullptr, nullptr);
-		return out;
-	}
-
-	inline std::string WideToUTF8(const std::wstring& ws)
-	{
-		return WideToUTF8(ws.c_str());
-	}
-
-	inline std::wstring UTF8ToWide(const std::string& s)
-	{
-		if (s.empty()) return std::wstring();
-
-		int len = MultiByteToWideChar(CP_UTF8, 0, s.c_str(), -1, nullptr, 0);
-		if (len <= 1) return std::wstring();
-
-		std::wstring out(len - 1, 0);
-		MultiByteToWideChar(CP_UTF8, 0, s.c_str(), -1, out.data(), len);
 		return out;
 	}
 }
