+++ libcommon/libcommon.vcxproj
@@ -103,11 +103,11 @@
     <ClCompile Include="src\bus\cartridge_menu.cpp" />
     <ClCompile Include="src\util\settings.cpp" />
     <ClCompile Include="src\util\filesystem.cpp" />
-    <ClCompile Include="src\util\winapi.cpp" />
     <ClCompile Include="src\util\DialogOps.cpp" />
     <ClCompile Include="src\util\Fileops.cpp" />
     <ClCompile Include="src\util\logger.cpp" />
     <ClCompile Include="src\util\fileutil.cpp" />
+    <ClCompile Include="src\util\coreutil.cpp" />
     <ClCompile Include="src\devices\cloud9.cpp" />
   </ItemGroup>
   <ItemGroup>
@@ -127,10 +127,11 @@
     <ClInclude Include="include\vcc\util\limits.h" />
     <ClInclude Include="include\vcc\util\logger.h" />
     <ClInclude Include="include\vcc\util\fileutil.h" />
+    <ClInclude Include="include\vcc\util\textutil.h" />
+    <ClInclude Include="include\vcc\util\coreutil.h" />
     <ClInclude Include="include\vcc\util\settings.h" />
     <ClInclude Include="include\vcc\util\critical_section.h" />
     <ClInclude Include="include\vcc\util\filesystem.h" />
-    <ClInclude Include="include\vcc\util\winapi.h" />
     <ClInclude Include="include\vcc\devices\cloud9.h" />
   </ItemGroup>
   <Import Project="$(VCTargetsPath)\Microsoft.Cpp.targets" />
