+++ libcommon/libcommon.vcxproj.filters
@@ -34,9 +34,6 @@
     <ClCompile Include="src\core\logger.cpp">
       <Filter>core</Filter>
     </ClCompile>
-    <ClCompile Include="src\core\winapi.cpp">
-      <Filter>core</Filter>
-    </ClCompile>
   </ItemGroup>
   <ItemGroup>
     <ClInclude Include="include\vcc\bus\basic_cartridge.h">
@@ -94,9 +91,6 @@
       <Filter>core</Filter>
     </ClInclude>
     <ClInclude Include="include\vcc\core\logger.h">
-      <Filter>core</Filter>
-    </ClInclude>
-    <ClInclude Include="include\vcc\core\winapi.h">
       <Filter>core</Filter>
     </ClInclude>
   </ItemGroup>
